requirements.txt
fastapi==0.128.8
uvicorn==0.39.0
openpyxl==3.1.5
python-multipart==0.0.20
aiofiles==25.1.0
python-dotenv==1.2.1
